requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-cors==0.4.0
